實打實大蘇打

+++ modules/game_help/PathMesh3D/extrude_profiles/path_extrude_profile_rect.cpp
@@ -20,7 +20,7 @@ Array PathExtrudeProfileRect::_generate_cross_section() {
 		norms.push_back(Vector2(0.0, -1.0));
 	}
 
-	for (uint64_t idx = 0; idx < subdivisions.x; ++idx) {
+	for (int32_t idx = 0; idx < subdivisions.x; ++idx) {
 		x += width_subdiv;
 		cs.push_back(Vector2(x, y));
 		norms.push_back(Vector2(0.0, -1.0));
@@ -37,7 +37,7 @@ Array PathExtrudeProfileRect::_generate_cross_section() {
 		norms.push_back(Vector2(1.0, 0.0));
 	}
 
-	for (uint64_t idx = 0; idx < subdivisions.y; ++idx) {
+	for (int32_t idx = 0; idx < subdivisions.y; ++idx) {
 		y += height_subdiv;
 		cs.push_back(Vector2(x, y));
 		norms.push_back(Vector2(1.0, 0.0));
@@ -54,7 +54,7 @@ Array PathExtrudeProfileRect::_generate_cross_section() {
 		norms.push_back(Vector2(0.0, 1.0));
 	}
 
-	for (uint64_t idx = 0; idx < subdivisions.x; ++idx) {
+	for (int32_t idx = 0; idx < subdivisions.x; ++idx) {
 		x -= width_subdiv;
 		cs.push_back(Vector2(x, y));
 		norms.push_back(Vector2(0.0, 1.0));
